falcon/requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
chromedriver-autoinstaller==0.6.4
click==8.2.1
curl_cffi==0.11.4
deep-translator==1.11.4
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.115.14
finnhub-python==2.4.23
Flask==3.1.1
flask-cors==6.0.1
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lxml==6.0.0
MarkupSafe==3.0.2
MouseInfo==0.1.3
multitasking==0.0.11
narwhals==1.44.0
numpy==2.2.6
openai==1.90.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
pdf2image==1.17.0
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
playwright==1.53.0
plotly==6.2.0
protobuf==6.31.1
pyarrow==20.0.0
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMuPDF==1.26.1
PyMySQL==1.1.1
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-Quartz==11.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytweening==1.2.0
pytz==2025.2
referencing==0.36.2
requests==2.32.4
requests-futures==1.0.2
rpds-py==0.25.1
rubicon-objc==0.5.1
scikit-learn==1.6.1
scipy==1.15.3
selenium==4.34.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
starlette==0.46.2
streamlit==1.46.0
ta==0.11.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.35.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yahooquery==2.4.1
yfinance==0.2.63

# fastapi==0.115.14
# uvicorn==0.35.0
# pydantic==2.11.7

# requests==2.32.4
# openai==1.90.0
# finnhub-python==2.4.23
# yfinance==0.2.63
# pandas==2.3.0
# PyMuPDF==1.26.1

# playwright==1.53.0
# selenium==4.34.0
# playwright install chromium
# pip install webdriver-manager